+++ src/basic/hooks/useCouponts.ts
@@ -1,0 +1,29 @@
+// TODO: 쿠폰 관리 Hook
+// 힌트:
+// 1. 쿠폰 목록 상태 관리 (localStorage 연동 고려)
+// 2. 쿠폰 추가/삭제
+//
+// 반환할 값:
+// - coupons: 쿠폰 배열
+// - addCoupon: 새 쿠폰 추가
+// - removeCoupon: 쿠폰 삭제
+
+import { useState } from 'react';
+
+import { initialCoupons } from '../../basic/constants';
+import { Coupon } from '../../types';
+
+export function useCoupons() {
+  const [coupons, setCoupons] = useState<Coupon[]>(() => {
+    const saved = localStorage.getItem('coupons');
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return initialCoupons;
+      }
+    }
+    return initialCoupons;
+  });
+  return { coupons, setCoupons };
+}
